requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
supabase-py>=2.4.0
